back-pjt/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-allauth==0.61.1
dj-rest-auth==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
Pillow==11.2.1
numpy==1.24.3
elasticsearch==8.12.1
scikit-learn==1.3.2
django-health-check==3.18.2
langchain==0.1.16
langchain-community==0.0.38
langchain-core==0.1.53
langchain-openai==0.1.5
langchain-text-splitters==0.0.2
pandas==2.2.3
scipy==1.15.2
matplotlib==3.10.0